% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_assign.R
\name{ntile_label}
\alias{ntile_label}
\title{Rank observations in n groups, with labels}
\usage{
ntile_label(var, n, digits = 0)
}
\arguments{
\item{var}{The variable wished to be ntile_label}

\item{n}{rank the variable in n groups}

\item{digits}{How many digits to include in the label}
}
\value{
A ordered factor vector of each n group. The value has the form of [min(n_i) - max(n_i)]
}
\description{
Rank observations in n groups, with labels
}
\details{
n_tile_label is very similar to ntile from dplyr. But n_tile_label creates
the n groups and then labels them. For each group i, the value of the ntile_label is 
[min(i) - max(i)].
}
\examples{
ntile_label(var = diamonds$price, n = 10)
diamantes<- diamantes \%>\% mutate(price_deciles = ntile_label(price, 10))
}
