% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCT.R
\docType{package}
\name{RCT}
\alias{RCT}
\title{RCT: A package for assigning and evaluating Randomized Control Trials}
\description{
The RCT provides three important group of functions:
a) functions for pre-processing the design of the RCT (tau_min, summary_statistics, ntile_label)
b) Functions for assigning treatment status and checking for balances (treatment_assign, balance_table, balance_regression)
c) Function for evaluating the impact of the RCT
}
